// src/contract/model/stationDetailsDtoApiResponse.ts
/**
 * EV Rental API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { StationDetailsDto } from './stationDetailsDto';

export interface StationDetailsDtoApiResponse {
  isSuccess?: boolean;
  data?: StationDetailsDto;
  message?: string | null;
}