// src/contract/model/stationDtoPagedResultApiResponse.ts
/**
 * EV Rental API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { StationDtoPagedResult } from './stationDtoPagedResult';

export interface StationDtoPagedResultApiResponse {
  isSuccess?: boolean;
  data?: StationDtoPagedResult;
  message?: string | null;
}